refactor(windows console): move to external file
